requirements.txt
Flask==2.0.1
numpy==1.22.0
tensorflow-lite==2.8.0
Pillow==9.0.0
gunicorn==20.1.0